fix: Handle EOF correctly on file read
refs: RUM-11629

This PR fixes a design flaw in how we read and process batch files for
upload.

 ### The problem: EOF isn't signalled until we read beyond EOF

Essentially, I made a blunder by misunderstanding a standard detail of
filesystem APIs. I wrote all this code with the assumption that when you
read the last byte of a file, the filesystem will report EOF.

In actuality, the filesystem does not report EOF until you attempt to
read past the end of the file. D'oh.

Unit tests have been working fine because I wrote the mock filesystem
implementation to behave according to the incorrect EOF behavior! I
noticed that the `std::filesystem` implementation behaves differently
while running live benchmarks.

 ### How this design flaw affects the upload thread

The code that reads TLV blocks from batch files expects that when it
reads the final block, an `eof` flag will be set: after processing the
block, it will stop reading in response to that `eof`.

In actuality, when we've read the last block from a file, we don't yet
know whether there are more blocks to read. Our reads are always aligned
to TLV block boundaries: we read the six-byte TLV header to determine
the size of the value, and then we read exactly that number of bytes. So
under normal circumstances, our last read leaves us at the end of the
file exactly, meaning the EOF bit isn't set.

As a result, the current implementation assumes there's another block in
the file, attempts to read it, fails to read a header, and treats the
lack of a well-formed header (where a block was expected) as
`TLVBlockReadResultType::Malformed`. As a result, it aborts the upload
and deletes the batch.

What we actually need to do is continue attempting to read TLV blocks
until we read 0 bytes: this is the correct criteria for deciding that
we've reaches the end of the file and have read all blocks successfully.

 ### Existing (flawed) implementation details

File reads happen across a few different layers, so that 1.) the
filesystem implementation is abstracted, 2.) the TLV-format business
logic is isolated and unit-testable, and 3.) the feature implementation
is presented with a straightforward iterator-ish interface rather than
needing to bother with the details of either filesystem or TLV decoding.

From low-level to high-level, here's the pertinent code.

`filesystem.hpp` defines the filesystem interface:

- `FileReadResult` is `struct { size_t num_bytes_read; bool eof; }`
- `IFileReader::Read` returns `expected<FileReadResult, FilesystemError>`

`tlv.hpp` defines the business logic for decoding TLV blocks given an
`IFileReader`:

- `TLVBlockReadResultType` is either `Success` or some kind of error
- `TLVBlockReadResult` is `struct { TLVBlockReadResultType type; TLVBlockHeader header; bool eof; }`
- `ReadTLVBlock` returns `TLVBlockReadResult`

`feature.hpp` defines the high-level interface used by feature
implementations to consume and process TLV blocks during upload:

- `TLVBlock` is `struct { TLVBlockType type; Block data; bool eof; }`
- `BatchReader::ReadNext` returns `expected<TLVBlock, BatchReadError>`

Finally, `TLVBatchWriter` (which streams event blocks from a file to an
HTTP request body) checks `TLVBlock::eof` after each read in order to
determine whether it should stop reading.

This old (and mistaken) approach of handling EOF after reading a valid
block required propagating the `eof` flag all the way through this
block-processing pipeline, which always felt a little gross.

 ### Corrected implementation

Here are the fixes in this PR:

- `FileReadResult` is removed
- `IFileReader::Read` returns `expected<size_t, FilesystemError>`
    - a value of `0` means we read zero bytes: when we've provided the
      function with a valid buffer, 0 (or any size less than the buffer
      size, really) is a clear indication of EOF
    - in `filesystem_std.cpp`, we handle the `eof` bit properly: if we
      read 0 bytes from `std::ifstream` and we're at EOF, we return a
      valid result; if we read 0 bytes and we're _not_ at EOF, we return
      an unexpected `FilesystemError::Failed`
- `TLVBlockReadResultType` is now either `Success`, or `EndOfFile`, or
  some kind of error
    - `Success` indicates that we read a valid block
    - `EndOfFile` indicates no more blocks; this is only the case when
      we read exactly 0 bytes when looking for the next 6-byte header
    - `IOError` and `ReadError` are still propagated from errors
      reported by the filesystem
    - `Malformed` is still returned if we hit EOF while reading (e.g.
      file contains an incomplete header or a partially-written block)
- `TLVBlock` is now just `struct { TLVBlockType type; Block data; }`
- `BatchReader::ReadNext` now returns `expected<optional<TLVBlock>, BatchReadError>`
- `TLVBatchWriter` now handles `std::nullopt` in place of a block as "we
  reached the end of the file successfully and there are no more blocks"

Fixing this issue required touching several different parts of the code,
which could have been quite gnarly... but luckily we have good test
coverage, so it was a pretty quick refactor!

The mock filesystem implementation has been updated to simulate the
correct read behavior, and tests are updated to clarify the expected
behavior re: reads at the EOF boundary.

Benchmarks now upload logging batches reliably.