chore: Add support for sanitizers (ASan and UBSan)
refs: RUM-7207, RUM-11498

The PR updates the CMake config to include a `DD_ENABLE_SANITIZERS`, on
by default in development builds, that configures all targets to build
and link (on supported platforms) with address sanitizer (ASan) and
undefined-behavior sanitizer (UBSan) enabled.

As a result, if any memory bugs or undefined behavior are detected while
running unit tests (or any other binary), execution will halt and the
sanitizer will print a bunch of useful debugging information.

Running tests with sanitizers enabled revealed a couple of minor issues,
which are fixed in this PR:

 ### UBSan: alignment in allocation-tracking implementation

When using `WITH_DATADOG_ALLOCATION_TRACKING`, the custom operator-new
implementation wasn't handling alignment conservatively enough, in ways
that led to undefined behavior.

Some common values MUST have 16-byte alignment, particularly on arm64
macOS, such that accessing them from a non-16-byte-aligned address is
UB. However, `alignof(std::max_align_t)` is still 8.

When allocation tracking is enabled, we use a custom `backend_malloc`
implementation, and it places an `AllocHeader` on the heap immediately
before the buffer we give back to the the caller.  `AllocHeader` was
previously `alignas(std::max_align_t)`, which caused the address
returned by `backend_malloc` to be 8-byte-aligned in places where the
compiler was expecting 16-byte alignment. Forcing the header to be
`alignas(16)` resolves the UB and makes our allocation-tracking
implementation consistent with the default behavior of `operator new` on
arm64 macOS.

 ### ASan: Reading beyond end of buffer in tests

A couple of tests in `feature_test.cpp` validate code that reuses
buffer space, doing something like this:

 - Write 8 bytes, `aaaaaaaa`, to a buffer
 - Check that the buffer holds `aaaaaaaa`
 - Write 4 bytes, `bbbb`, to the same buffer
 - Check that the buffer holds `bbbbaaaa`

ASan doesn't like this. And in fact, it's not great test design: it's
not our business to know what's left hanging around in a `std::vector`'s
underlying buffer beyond its current size. The tests don't suffer from
removing these checks and sticking to simply validating the observable
behavior.