chore: Add clang-format support to CMake, restore format targets
refs: RUM-11498, RUM-7207

This PR makes clang-format a part of the CMake build, such that you
don't need to externally install `clang-format`: if you run a CMake
build with clang-format support enabled, CMake will install the
appropriate clang-format binary for you.

clang-format is distributed with LLVM, and each major release has
differences in behavior: e.g. clang-format-15 and clang-format-20 will
give you different results for the same code with the same set of rules.
Therefore, it's important to establish a canonical clang-format version.
This PR establishes clang-format 20.1.8 as the standard version for
formatting this repo, irrespective of the build toolchain used to
compile one build or another.

If you don't want to bother with clang-format, you can simply leave
`DD_ENABLE_CLANG_FORMAT` disabled and your build won't do anything
special. If it _is_ enabled, then:

- If you already have `clang-format` installed, and CMake can find it in
  the path, and `clang-format --version` reports
  `clang-format version 20.1.8`: then you're good to go; CMake will use
  this version.

- If you don't have `clang-format` installed, or if the `clang-format`
  found in the system path is not version 20.1.8: CMake will download
  the appropriate LLVM release archive for your platform, extract
  `bin/clang-format` to a `tools/bin` directory within this repo, and
  then use that version thereafter.

Downloading and unzipping the entire LLVM release takes a couple
minutes, so we wouldn't want to incur that hit in CI builds. The CI job
that validates formatting up-front is the only job that needs to use
`DD_ENABLE_CLANG_FORMAT`, and it can be configured to use Clang 20 as
its toolchain of choice, so that it already has the requisite
`clang-format` binary available.

This PR also adds `format` and `check-format` targets to the build.
Running `cmake --build <build-dir> --target format` will reformat all
SDK source files according to the rules in `.clang-format`; running the
`check-format` target will validate that all source files are already
formatted according to those rules.