fix: Don't download clang-format by default; require explicit option
Follow-up to #44.

If `DD_ENABLE_CLANG_FORMAT` is set, `DD_DEVELOPMENT_ALLOW_AUTO_INSTALL`
must also be set in order to auto-download clang-format v20 to
`tools/bin/`. If auto-install is not enabled, the script will print a
warning and proceed with the build, without configuring format targets.

Additionally, when the script _does_ auto-download an LLVM release in
order to install clang-format, it now verifies the SHA256 of the
downloaded file. If checksum verification fails, the build halts.