chore: Support auto-installing both clang-format and clang-tidy
refs: RUM-11498

In order to be operating within the golden-path development environment,
you need both clang-format _and_ clang-tidy at a specific version. This
PR updates the auto-download logic in the CMake build so that both tools
will be installed (now to a more-narrowly-scoped `llvm-bin/` directory)
from the appropriate LLVM release archive iff either tool is not present
in the PATH and `DD_DEVELOPMENT_ALLOW_AUTO_INSTALL` is enabled.