fix: In C++ API, use DATADOG_API at function level on PIMPL types
refs: RUM-11377, RUM-11376

Declaring `__declspec(dllexport)` at the class level tags every member
of the class as exported, even private members. MSVC warns when a data
member of an exported type refers to a type that is not itself exported,
regardless of access specifier:

```
warning C4251: 'datadog::Core::_impl':
  'std::unique_ptr<datadog::impl::Core>'
    needs to have dll-interface to be used by clients of 'datadog::Core'
```

We could circumvent this by suppressing the warning and continuing to
use `DATADOG_API` on API types that use PIMPL, since `_impl` is truly
private and should never be accessed publicly anyway.

However, I would contend that it's better to explicitly tag each
exported member function. This way, private data and member functions
don't get accidentally exported, nor do auto-generated rule-of-five
functions. If we want copying/moving to be supported as part of the
public API, we need to be explicit about it.

As of this PR:

- All user-facing functions in the C++ API are still exported
- `Core`, `Logging`, and `Logger` no longer export an implicit-default
  public constructor
- `Core`, `Logging`, and `Logger` now explicitly delete copy/move
  operations
- `Core`, `Logging`, and `Logger` now explicitly export the default
   destructor (which handles pimpl cleanup via smart pointer dtor)
- Constructors now use proper C++ initialization semantics (i.e. the
  impl value is moved in via the ctor, not assigned to the member after
  construction)
- `CoreConfig` gets the same treatment as the above types, although it's
  explicitly copyable and movable (using default implementations of all
  special member functions)

Note that:

- Destructors have to be defined (even as `= default`) in the .cpp file,
  where the underlying `impl` type managed by the smart pointer is
  complete
- Constructors still need to be public if we want to use them with
  `std::make_shared` etc.; adding a `PrivateCtorTag` argument ensures
  that they still can't be used publicly
    - We could keep the constructor fully private and use
      `std::shared_ptr(new Core())`, but this would require a separate
      (and potentially non-cache-adjacent) allocation for the control
      block each time we created an API resource
- These types (`datadog::Logging` et al.) are just wrappers, and we pass
  them to callers at the API boundary via `std::shared_ptr`, so there's
  no need for the API type itself to be copyable or movable: the user
  should be copying/moving `std::shared_ptr<datadog::Core>`, never
  dealing with a naked `datadog::Core`.