fix: Replace dd_logger_config_create with dd_logger_config_init
refs: RUM-11496, RUM-11314

This PR updates the C API to expose a non-opaque struct for
`dd_logger_config_t`. This change requires callers to provide their own
`dd_logger_config_t` value to the API, rather than requiring that the
API heap-allocate every logger config value.

Instead of `dd_logger_config_create(void)`, we now have
`dd_logger_config_init(dd_logger_config_t*)`.
`dd_logger_config_destroy(dd_logger_config_t*)` has been removed.

There are a few design tradeoffs to consider here.

 ### String storage

As of this PR, `dd_logger_config_t` stores its `service` and `name`
values in local, fixed-size buffers.

We could accept `const char*` in order to permit strings of arbitrary
length: this would work fine for string literals with static storage, or
for stack-local buffers that persist throughout the lifetime of the
config.

However, the lifetime of a `dd_logger_config_t` is managed by the user,
and they may keep a value around persistently for reuse. It's safer if
we store a copy of the string data passed to `dd_logger_config_set_service`
and `dd_logger_config_set_name`.

This means we have to set a reasonable upper limit on the length of
these values: this PR settles on `char[127 + 1]` for the service name
and `char[63 + 1]` for the logger name.

 ### Truncating vs rejecting long strings

As a consequence of the above, we need to decide how to handle an input
value that exceeds the max string length. If a service or logger name
exceeds the maximum length that we can store, we _could_ reject it
outright, in which case we'd want our API to have proper error handling
on these functions (e.g. we return `dd_error_t` and expect the caller to
check for success after setting a name).

Since we tend to err on the side of "just roll with it" when it comes to
handling non-critical errors in the API, and since this is an
implementation detail of the C API, and since it's simplest to implement
this way in a v0/v1 release: as of this PR, the API will just truncate
long strings and carry on.

This _does_ mean that if you create two loggers with these distinct
names:

- `this-is-the-logger-for-my-application-and-this-logger-handles-initialization-events`
- `this-is-the-logger-for-my-application-and-this-logger-handles-input-events`

Then the resulting events from both loggers will be tagged with the same
truncated name:

- `{"logger":"this-is-the-logger-for-my-application-and-this-logger-handles-i"}`

 ### Versioning

I've also tossed in a `uint32_t version` member as the struct's first
member. This value serves two purposes:

1. It's a magic value that we can use to determine whether a config
   struct has been properly initialized: i.e. if the user didn't call
   `dd_logger_config_init` and just left parts of the struct
   uninitialized (or zero-filled), we can detect that and fall back to
   the default config.

2. It gives us a means of detecting and handling ABI compatibility
   issues, in the event that we add new fields to the struct and we want
   to be defensive about backward-compatibility.