fix: Make core+logging APIs safe to call unconditionally
refs: RUM-11758

In the core and logging APIs, there are currently cases where we return
`nullptr` in response to:

1. Unacceptable user-supplied arguments
   _(e.g. passing an invalid `Core` ptr to `Logging::Register`)_

2. Use of the API in an unsupported state
   _(e.g. attempting to register a duplicate feature)_

3. Internal failures
   _(e.g. failure to initialize filesystem or HTTP subsystem)_

This PR updates all existing APIs to ensure that the SDK will never
return a value that is unsafe to use. e.g. you can safely call
`Logging::Register(core)->CreateLogger()->Info("hello")` if you really
want to, and if `core` is not actually valid, nothing bad will happen.

We actually already get the intended behavior for free in the C API,
where there's no such thing as a member function. Returning `nullptr` is
fine, because every API function checks for `nullptr` and simply does
nothing.

In C++, the API layer deals in objects (e.g. `datadog::Logging`) that
wrap underlying implementation-layer objects (e.g. `impl::Logging`). So
we can unconditionally return the API wrapper object, and implement
every member function to handle `if (!_impl)` with a no-op.

Code changes:

- `Core::Create` now returns a valid wrapper (with a null impl) instead
  of returning `nullptr`
- `Core::Create` now rejects the SDK config if it does not contain a
  valid `client_token`
- All `Core` methods check `_impl != nullptr`
- `core_cpp_api_test.cpp` validates usage of no-op `Core`
- `core_c_api_test.cpp` validates null checks in `dd_core` API
- `Logging` and `Logger` get the same treatment: if required args or
  wrapped impl objects are null, we return a valid API object that wraps
  a `null` impl
- `logging_cpp_api_test.cpp` validates usage of no-op `Logging` and
  `Logger`

Note that if shared-library builds are eventually modularized with
separate binaries per-feature (see [RUM-11669](https://datadoghq.atlassian.net/browse/RUM-11669)),
we'll need to answer some additional design questions re: no-op
implementations - namely, whether we want it to be possible to call APIs
for features that are not even linked into the build. For this PR (and
for RUM-11758 as a whole), we make no attempt to tangle with those
questions.