chore: Update README and CONTRIBUTING docs, remove build.sh
refs: RUM-11205

One last thing before we merge to main: I wanted to make a quick pass at
revising the documentation in `README.md` and `CONTRIBUTING.md`. The
former now includes instructions for integrating the SDK into a CMake
build; the latter now has more detailed workflow instructions for SDK
development.

Also, `build.sh` is vestigial: just using CMake is the way to go. If we
end up wanting to layer a bit more DX convenience on top of manual
`cmake` commands, it seems that CMake Presets are the standard way to
handle that.

Re: reviews from the documentation team - if we feel it's warranted, we
should explicitly request a review / copy-editing pass (with the
customary two weeks' lead time) before we release. From recent PRs I've
seen in other repos, I'm led to believe that this is the preferred
approach (as opposed to automatically tagging the docs team on every
change via CODEOWNERS).