feat: Set a unique DD-REQUEST-ID header on all outgoing HTTP requests
refs: RUM-11357

This PR updates the core's HTTP-request-building logic to use the new
`UUID` type. Each time we prepare an HTTP request in the upload thread,
we use `UUID::Random()` to generate a value to use as that request's
`DD-REQUEST-ID` header.

To avoid allocation, we use a new `UUID::ToBuffer()` function to
serialize the value directly to the header string.

Resolves [RUM-11357](https://datadoghq.atlassian.net/browse/RUM-11357).