feat: Enrich log events with RUM context
refs: RUM-11951

As of this PR, by default (unless explicitly disabled in the logger
config), log events will have the following attributes automatically
added to them, provided that RUM is in use:

- `application_id`: unconditionally, as long as the configured RUM
   application ID is a valid, nonzero UUID
- `session_id`: if there's an active session
- `view.id`: if there's an active view
- `user_action.id`: if there's an ongoing action

These values are provided by a new `RumFeatureContext` struct, which is
stored as a member of `CoreContext`. Note that while the mobile SDKs use
dynamic key-value lookups to store feature-specific context values, here
we just make them plain old members. This seems like the simplest
approach to me, given that:

- The set of feature-specific context types is small and is known at
  compile-time
- `CoreContext` is copied frequently, and we want those copies to be as
  cheap as possible

Note that the RUM feature implementation doesn't yet populate these
`RumFeatureContext` fields, but the RUM-enrichment functionality added
to logging in this PR is fully unit-tested.

Additionally:

- The shared surface area between `Logger` and `Logging` was getting
  pretty gnarly; I wrapped all the shared state into a `LoggerState`
  struct to more clearly delineate what state is internal to the Logger
  and what state is used in `OnLoggerEmit`, and to make the
  `LogEventCallback` signature less of a nightmare

- In the tests for the logging C API, we were missing null-safety tests
  for `logger_config_*`; coverage has been added

- We previously set logger name and version attributes as
  `{"logger":{"name":"foo","version":"0.0.0"}}`; this has been corrected
  to `{"logger.name":"foo","logger.version":"0.0.0"}`

- A new test support class, `FeatureTest`, facilitates unit-testing of
  feature implementations, injecting a test-controlled `CoreContext` and
  capturing all generated events for examination