feat: Store RUM Application ID as UUID; reject non-UUID values
refs: RUM-11367

This is a minor change to support the addition of RUM sessions.

Within our SDKs, RUM uniquely identifies sessions, views, actions, etc.
by generating UUIDs. RUM also maintains a value that identifies the
configured RUM Application: this "RUM Application ID" value is provided
by the user on SDK init.

Our existing SDKs (i.e. Android and iOS) accept any arbitrary string
value: as long as the RUM Application ID is non-empty, the SDK will
faithfully use it, tag events with it, etc.

From what I can gather, though, RUM Application ID values generated by
the Datadog platform are always UUIDs. In the C++ SDK, wherever we can
store UUIDs instead of strings, the code will generally be more
straightforward as a result. Compared to `std::string`, a UUID value has
a known, fixed size, and is therefore trivially copyable - it doesn't
need to allocate heap memory, it can't throw exceptions due to OOM, etc.

This PR updates the C++ SDK to require a nonzero UUID for the
user-provided RUM Application ID value, so that we can store the value
as a UUID and treat it as such throughout the lifetime of the SDK,
without any intermediate string <-> UUID conversions.

If the user attempts to initialize the RUM feature with an application
ID value that isn't a nonzero UUID, the SDK will return a no-op RUM
interface, as we can reasonably assume that a non-UUID application ID is
not valid.

If there are legitimate use cases for non-UUID application ID values,
please let me know.