feat: Establish basic RUM types
refs: RUM-11367

This PR establishes the core concepts used to model RUM application
state. In a nutshell:

- RUM models application state as a hierarchy rooted at the
  **Application**, which has one active **Session**, which has multiple
  **Views** (one of which may be active at any given time), in which any
  number of **Actions** and **Resources** can be recorded.

- A RUM **Scope** represents one of these points in the hierarchy:
  `RumApplicationScope` is the root, `RumSessionScope` sits beneath it
  and owns a set of `RumViewScope`s, etc.

- A RUM **Command** is an event that may be propagated through the tree
  in order to precipitate a state change. Each **Scope** can _process_ a
  **Command**, choosing whether to propagate it to child scopes, and the
  result of _processing_ is that the scope either _remains open_ or is
  _closed_.

- A RUM **Context** is an immutable snapshot representing the current
  state of the scope tree. Each **Scope** can _populate_ a **Context**,
  describing what it knows about itself and any parent scopes.

The implementation of `RumApplicationScope` and `RumSessionScope` will
follow in a separate PR. This PR is mostly meant to validate:

1. The broad-strokes design conventions that I've lifted from the mobile
   SDKs' RUM implementations, and

2. How I've chosen to implement those designs in C++

 ### Externally-defined RUM data types

[A recent PR](https://github.com/DataDog/dd-sdk-cpp/pull/74) introduced
`core/feature_types` as a place to define the data types that describe a
specific feature, but that are meant for external consumption throughout
the SDK, and thus belong in the core.

In `core/feature_types/rum.cpp`, we now define `RumSessionPrecondition`.
This string enumeration is part of the JSON schema defined in
[rum-events-format](https://github.com/DataDog/rum-events-format/).

We're not yet generating any RUM events, and `RumSessionPrecondition` is
the only external type that we need at the moment, so it's easy enough
to maintain C++ types (that match the JSON schema) by hand.

In the short-term, given that we want to get basic RUM functionality
working ASAP, that's probably the approach we should stick with.
However, the maintenance burden _will_ scale poorly as we support new
RUM event types and work on other features.

Longer-term, we may want to spend some time (maybe a couple of weeks?)
to come up with a better solution. From where I stand, we would want to:

- Decouple JSON serialization from the Attribute API
- Support structured payload types (e.g. `RumViewEvent` with named
  fields, not just `Object::Attribute()` with arbitrary string keys)
- Come up with tooling (similar to what the iOS and Android SDKs have)
  to generate C++ types from the schemas in rum-events-format

 ### RUM implementation

The remainder of the changes are internal to `features/rum`, and they
essentially implement the concepts described above.

Note that `RumScope` and `RumContextProvider` aren't actually used as
abstract base classes; they're just duck-typed interfaces that are
documented in the source for clarity. (As far as I can tell, we don't
actually need runtime polymorphism for scopes, as we're always
traversing the hierarchy in a well-defined order, with full compile-time
knowledge of the types we need to deal with at each level of the tree.)

 ### Test coverage

- `rum/context_test.cpp` tests the conversion from `RumContext` (an
  internal snapshot of the scope tree) to `RumFeatureContext` (a more
  streamlined description of the current RUM context for use in
  enriching other features' events)

- `rum/scope_test.cpp` tests the session-sampling-rate logic that's
  included in `RumScopeDependencies`