fix: Restart sessions more permissively after explicit stop
refs: RUM-12259

This PR fixes a slight error in the session lifecycle logic implemented
in PR #78.

 ### The bug

Basically, if you explicitly stop a session by calling `StopSession()`,
the `RumApplicationScope` will be left without an active session. The
next time it processes any command, it has to decide whether to create a
new session to handle that command (or else simply drop the command).

Currently, it will only create a new session if the command represents a
user interaction. This is incorrect: in this situation, we want to
create a new session in response to pretty much any command. (The rules
for determining whether we transfer existing _views_ from the previous
session are more complicated, but that's irrelevant here.)

 ### The fix

This PR fixes the logic, ensuring that `RumApplicationScope` will create
a new `RumSessionScope` to handle any command received after an explicit
`StopSession` call - the only exceptions are `StopSession` itself (which
should rightly be ignored if already stopped) and internal commands like
`SDKInit` that don't correspond to _any_ application-initiated event.

As before, a user interaction command still serves its intended purpose
of refreshing the inactivity timeout in the session scope.

 ### Command type traits

While we're at it, this PR also refactors `is_user_interaction` so that
it's no longer a per-instance member variable: this flag is now defined
as intrinsic to the command type, and it can be accessed at runtime by
calling `command.IsUserInteraction()`

A similar flag, `should_create_new_session_after_explicit_stop`, is used
to denote whether a particular type of command should result in a new
session being created in the no-active-session case.