feat: Implement RUM View API
refs: RUM-11368

This PR exposes API functions for starting and stopping RUM views:

- C: `dd_rum_start_view()` / `dd_rum_stop_view()`
- C++: `datadog::Rum::StartView()` / `datadog::Rum::StopView()`

This PR also adds some API tests that exercise basic RUM view usage
end-to-end, validating that view events are produced as expected.

Note, however, that we only produce `{"placeholder-for":"view"}`
payloads as of this PR. Properly-formatted RUM view events will be
tackled separately in RUM-12321, which may depend on RUM-12171 and
RUM-12172.

Also, custom attributes are not fully supported, and there is no API for
modifying view attributes yet: this will be done in RUM-12322.