feat: Add API-layer timestamp types
refs: RUM-12174

This PR is a quick prerequisite for RUM-12174, which deals with how we
support JSON-serializing timestamp values.

We use timestamps internally throughout the SDK. These usages rely on
the `platform::Duration` and `platform::Timestamp` types, which are
defined in the "Clock" platform subsystem. These types are aliases for
`std::chrono` representations.

We _also_ expose some limited timestamp usage at the API layer, since
we support attribute values that are explicitly typed (and efficiently
stored) as timestamps. The Attribute API doesn't use
`platform::Timestamp` (as it's an implementation-layer detail): instead,
it accepts raw `uint64_t` nanoseconds counts, with some `std::chrono`
overloads in the C++ API for convenience.

The SDK should be consistent in how it handles timestamp values. Since
we expose timestamps at the API layer, and since they're simple,
fundamental data types with very little maintenance burden, I think it
makes sense to lift our canonical timestamp definitions up to the API
layer.

This PR moves the definition of `datadog::impl::platform::Timestamp` to
`datadog::Timestamp`, exposing it via the C++ API. At the same time, it
establishes `dd_timestamp_t`, which is simply a type alias for
`int64_t`, with a few initialization functions that clarify the expected
semantics (count of nanoseconds since the Unix epoch).

The API layer is only concerned with establishing the Datadog SDK's
canonical representation for timestamp values, and ensuring that users
can easily and correctly initialize values in that format. There's no
compelling reason to provide any other timestamp-related functionality
at present.