feat: Define Omissible<T> for optional struct fields
refs: RUM-12172

 ### Rationale

Most event types produced by the SDK have some optional fields: that is,
fields which we want to entirely omit from the JSON payload if they are
unchanged from the default/zero/null value.

For example, if our RUM context includes a valid application_id and
session_id but no view_id or action_id, we don't want to produce this:

```
{
  "rum": {
    "application_id": "9d71af28-73d4-4717-a12c-f2ca3855773d",
    "session_id": "68c4d4e1-a5c9-42a9-9021-51c485b48106",
    "view_id": "00000000-0000-0000-0000-000000000000",
    "action_id": "00000000-0000-0000-0000-000000000000"
  }
}
```

...rather, we'd prefer to produce a more compact payload:

```
{
  "rum": {
    "application_id": "9d71af28-73d4-4717-a12c-f2ca3855773d",
    "session_id": "68c4d4e1-a5c9-42a9-9021-51c485b48106"
  }
}
```

 ### Code changes

This PR adds `Omissible<T>`, which wraps an underlying value of type
`T`. When a struct containing a value of `Omissible<T>` is serialized as
a JSON object:

1. If `value == T{}`, it will be entirely omitted from the JSON object
2. If `value != T{}`, the value will be serialized normally as a JSON
   object property

This allows us to define:

```cpp
struct RumContext {
  Omissible<UUID> application_id;
  Omissible<UUID> session_id;
  Omissible<UUID> view_id;
  Omissible<UUID> action_id;
};

DATADOG_JSON_STRUCT(
  RumContext,
  DATADOG_JSON_FIELD(application_id),
  DATADOG_JSON_FIELD(session_id),
  DATADOG_JSON_FIELD(view_id),
  DATADOG_JSON_FIELD(action_id)
)
```

...to get the desired behavior outlined above.