fix: Allow indirect initialization via implicitly-convertible arg types
The previous commit is a great example of why C++ templates are great.
This commit is a great example of why C++ templates are terrible.

This change allows us to implicitly construct an `Omissible<T>` from a
value of any type `U` where `T` is implicitly constructible from `U`.

This ensures that we can initialize a member of type
`Omissible<std::string>` from `std::string_view`, `const char*`, etc.