fix: Ensure that log events do not include stale attribute values
refs: RUM-12499

When we merge user attributes prior to serializing a log event to JSON,
the destination object that contains the set of merged attributes is a
separate `Attribute` value that we reuse between events, to ensure that
we're not constantly allocating memory with each new event.

Previously, we were failing to properly reset the state of this object
on each usage. This meant that old attribute values continued to be
included in log events, even after those attributes had been deleted by
the user.

This PR adds test to cover this case, and it also fixes the bug.

Note that when we clear the object, we retain the underlying storage
held by the reusable `Attribute` (i.e. capacity in the `CowValue`'s
`std::vector<std::pair<std::string, Attribute>>`). Clearing the object
_does_ release any references to `CowValue`s held in that vector, but
each of those references is unlikely to be the last: they ought to still
be shared with the original `Attribute` they were copied from. In other
words: if we continually generate lots of log events with the same set
of custom attributes, reallocation should not be a common occurrence.