chore: Add nlohmann/json for better JSON validation in tests
refs: RUM-12511

 ### Changes in this PR

This PR adds `nlohmann/json` as a development-only dependency, solely
for the unit test target, so we can improve the subset of tests that
semantically validate JSON data produced by the SDK.

In implementation-layer tests that validate the SDK's JSON serialization
implementation, we still strictly validate the resulting values as
strings, with byte-for-byte comparison. The test-only helper function
used for that purpose is now called `RequireJsonLiteral`, for clarity.

In API layer tests that validate the events produced in response to a
set of API calls, though, we were previously using basic string
manipulation to parse and validate JSON values. As of this PR, we now
_actually_ parse and validate these values into `nlohmann::json` values,
allowing us to easily perform semantic validation, check for the
presence of specific attributes, etc.

This is a nice-to-have improvement for simple log events, but it's a
critical requirement for working with complex RUM events.

 ### Choice of JSON library

`nlohmann/json` is MIT-licensed, well-supported, written in C++11, has
no dependencies, and is implemented in a single header. While it doesn't
outperform `rapidjson` in terms of memory efficiency or raw speed, it
has an expressive and convenient-to-use modern C++ API. Since we're only
using this library in unit tests, and not in production, usability
prevails over speed.