feat: Add tools/validate-event-types
refs: RUM-12525

 ### Tool for externally validating RUM events from C++ unit tests

This PR adds a Python script in `tools/validate-event-types`. To quote
the description in `main.py`:

> Usage: `python3 tools/validate-event-types/main.py [rum-events-format-commit-ref]`
>
> Running this script will:
>
> 1. Prepare a local clone of the `rum-events-format` repo at the
>    specified revision (latest if no revision specified)
> 2. Use `python-jsonschema` to parse
>    `rum-events-format/schemas/rum-events-schema.json`
> 3. Find every invocation of the `RequireValidRumEvent` function in
>    `tests/impl/core/feature_types/rum_test.cpp`, and parse the second
>    operand, which must be a C++ raw-string literal containing a
>    complete JSON object literal
> 4. Use `python-jsonschema` to validate that JSON object literal
>    against the schema
>
> This validation process does not directly exercise the functionality
> of the C++ SDK. The C++ unit test is responsible for validating that
> the SDK produces the expected JSON payload exactly as shown in the
> `RequireValidRumEvent` call. This script separately validates that the
> expected JSON payload is indeed a valid RUM event according to the
> schema. If both hold true, we know that the C++ SDK produces valid RUM
> events within the extent of our test coverage.

 ### CMake target for invoking event validation

This PR also adds a new CMake build option, `DD_ENABLE_EVENT_VALIDATION`,
which is on by default in development builds (and off by default
otherwise). When enabled, we define a `validate-event-types` target, so
that locally, or in our own CI process, we can run:

- `cmake --build build --target validate-event-types`

...to verify that all RUM events used in the relevant unit tests are
valid according to the schema. The CMake target seamlessly resolves
Python 3.10+ (failing with a descriptive error if not found),
initializes a `.venv` directory, and installs dependencies prior to
running the script.

 ### Support for invoking event validation on `rum-events-format` PRs

In the CI pipeline for `rum-events-format`, we could also exercise these
tests by cloning `dd-sdk-cpp` and running:

```
cmake \
  -DDD_ENABLE_EVENT_VALIDATION=ON \
  -DEVENT_VALIDATION_COMMIT_REF=<branch-name-or-commit-sha> \
  . -B build && \
cmake --build build --target validate-event-types
```

We can probably hold off on making that change until `dd-sdk-cpp` is
public.

 ### Changes to our CI pipeline

This PR _does_ update the CI pipeline for `dd-sdk-cpp` such that we
always run event validation alongside unit tests.